+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 1.1.0 (1 Feb 2021)
+#### News
+- Add new StorageType: `.files`
+#### Changes
+- StorageKey type changed from `String` to `CustomStringConvertible`
+- Initializers were changed to static methods
+
 ## 1.0.0 (12 Dec 2021)
 Initial release